# Conflicts:
#	shesha-core-enterprise

+++ Boxfusion.Health.His.Admissions.Domain.Tests/AdmissionsTestBase.cs
@@ -22,6 +22,7 @@ using Boxfusion.Health.His.Admissions.Domain.Tests.Admissions;
 using Boxfusion.Health.His.Common.Diagnoses;
 using Boxfusion.Health.His.Common.ConditionIcdTenCodes;
 using Boxfusion.Health.His.Common.Admissions;
+using Boxfusion.Health.His.Common.BillingClassifications;
 
 namespace Boxfusion.Health.His.Admissions.Tests
 {
@@ -32,8 +33,12 @@ namespace Boxfusion.Health.His.Admissions.Tests
         protected IRepository<HisWard, Guid> _wardRepository;
         protected IRepository<Diagnosis, Guid> _diagnosisRepository;
         protected IRepository<HisConditionIcdTenCode, Guid> _conditionIcdTenCodeRepo;
+        protected IRepository<HospitalAdmission, Guid> _hospitalAdmissionRepository;
+        protected IRepository<BillingClassification, Guid> _billingClassificationRepository;
+        protected IRepository<HisHealthFacility, Guid> _hisHealthFacilityRepo;
         protected readonly DiagnosisManager _diagnosisManager;
         protected readonly ConditionIcdTenCodeManager _conditionIcdTenCodeManager;
+
         protected IUnitOfWorkManager _uowManager;
         protected ISessionProvider _sessionProvider;
 
@@ -44,6 +49,9 @@ namespace Boxfusion.Health.His.Admissions.Tests
             _wardRepository = Resolve<IRepository<HisWard, Guid>>();
             _diagnosisRepository = Resolve<IRepository<Diagnosis, Guid>>();
             _conditionIcdTenCodeRepo = Resolve<IRepository<HisConditionIcdTenCode, Guid>>();
+            _hospitalAdmissionRepository = Resolve<IRepository<HospitalAdmission, Guid>>();
+            _billingClassificationRepository = Resolve<IRepository<BillingClassification, Guid>>();
+            _hisHealthFacilityRepo = Resolve<IRepository<HisHealthFacility, Guid>>();
             _diagnosisManager = Resolve<DiagnosisManager>();
             _conditionIcdTenCodeManager = Resolve<ConditionIcdTenCodeManager>();
             _uowManager = Resolve<IUnitOfWorkManager>();
@@ -104,6 +112,21 @@ namespace Boxfusion.Health.His.Admissions.Tests
             return await _wardRepository.FirstOrDefaultAsync(a => a.Name == name);
         }
 
+        protected async Task<HospitalAdmission> GetTestData_HospitalAdmissionAsync(Guid id)
+        {
+            return await _hospitalAdmissionRepository.GetAsync(id);
+        }
+
+        protected async Task<BillingClassification> GetTestData_BillingClassificationAsync(Guid id)
+        {
+            return await _billingClassificationRepository.GetAsync(id);
+        }
+
+        protected async Task<HisHealthFacility> GetTestData_HisHealthFacilityAsync(Guid id)
+        {
+            return await _hisHealthFacilityRepo.GetAsync(id);
+        }
+
         protected async Task<HisPatient> CreateTestData_NewPatient(string name)
         {
             HisPatient patient;
@@ -140,7 +163,7 @@ namespace Boxfusion.Health.His.Admissions.Tests
 
 					MapToAdmissionEntity(admission, result);
                     var codes = await GetTestData_CodesList();
-                    await MakeADiagnosis(admission, codes, RefListEncounterDiagnosisRoles.AD, RefListAdmissionStatuses.admitted);
+                    await MakeADiagnosis(admission, codes, RefListEncounterDiagnosisRoles.AD, RefListWardAdmissionStatuses.admitted);
 
                     session.Flush();
                     await uow.CompleteAsync();
@@ -159,7 +182,7 @@ namespace Boxfusion.Health.His.Admissions.Tests
 		{
             admission.Id = result.Id;
             admission.WardAdmissionNumber = result.WardAdmissionNumber;
-            admission.AdmissionStatus = (RefListAdmissionStatuses?)result.AdmissionStatus;
+            admission.WardAdmissionStatus = (RefListWardAdmissionStatuses?)result.WardAdmissionStatus;
             admission.AdmissionType = (RefListAdmissionTypes?)result.AdmissionType;
 			admission.PartOf = new HospitalAdmission
 			{
@@ -167,7 +190,7 @@ namespace Boxfusion.Health.His.Admissions.Tests
 			};
 		}
 
-		private async Task MakeADiagnosis(WardAdmission wardAdmissionEntity, List<IcdTenCode> codes, RefListEncounterDiagnosisRoles use, RefListAdmissionStatuses status)
+		private async Task MakeADiagnosis(WardAdmission wardAdmissionEntity, List<IcdTenCode> codes, RefListEncounterDiagnosisRoles use, RefListWardAdmissionStatuses status)
         {
             var diagnosisEntity = await _diagnosisManager.AddNewDiagnosis<WardAdmission, Condition>(wardAdmissionEntity,
                                           (int)use, null,
@@ -185,7 +208,7 @@ namespace Boxfusion.Health.His.Admissions.Tests
                                                                         //(Optional) extra values required in ConditionIcdTenCode sub-entities
                                                                         async item =>
                                                                         {
-                                                                            item.AdmissionStatus = status;
+                                                                            item.WardAdmissionStatus = status;
                                                                         });
         }
 
@@ -241,7 +264,8 @@ namespace Boxfusion.Health.His.Admissions.Tests
         protected void CleanUpTestData_Patient(Guid patientId)
         {
             using var session = OpenSession();
-            var query = session.CreateSQLQuery("DELETE FROM Core_Persons WHERE Id = '" + patientId.ToString() + "'");
+            var query = session.CreateSQLQuery($"Update Fhir_ChargeItems set SubjectId = null where SubjectId = '{patientId}'" +
+                "DELETE FROM Core_Persons WHERE Id = '" + patientId.ToString() + "'");
             query.ExecuteUpdate();
 
             session.Flush();
@@ -253,12 +277,14 @@ namespace Boxfusion.Health.His.Admissions.Tests
 
             using var session = OpenSession();
             var query = session.CreateSQLQuery($"Update Fhir_Conditions set FhirEncounterId = null where FhirEncounterId = '{admission.Id}'" +
+                 $"Update His_BedOccupations set WardAdmissionId = null where WardAdmissionId = '{admission.Id}'" +
                 $"DELETE FROM Fhir_Encounters WHERE Id = '{admission.Id}'");
             query.ExecuteUpdate();
 
             if (includePartOf)
             {
                 query = session.CreateSQLQuery($"Update Fhir_Conditions set HospitalisationEncounterId = null where HospitalisationEncounterId = '{admission.PartOf.Id}'" +
+                    $"Update Fhir_ChargeItems set ContextEncounterId = null where ContextEncounterId = '{admission.PartOf.Id}'" +
                 $"DELETE FROM Fhir_Encounters WHERE Id = '{admission.PartOf.Id}'");
                 query.ExecuteUpdate();
             }
@@ -267,7 +293,7 @@ namespace Boxfusion.Health.His.Admissions.Tests
             {
                 diagnosisList.ForEach(diagnosis =>
                 {
-                    DeleteDiagnosisCombo(diagnosis);
+                    DeleteDiagnosisCombo(diagnosis, false);
                 });
             }
             else
@@ -278,14 +304,17 @@ namespace Boxfusion.Health.His.Admissions.Tests
             session.Flush();
         }
 
-		private void DeleteDiagnosisCombo(Diagnosis diagnosis)
+		protected void DeleteDiagnosisCombo(Diagnosis diagnosis, bool hasIcdTen = true)
 		{
             using var session = OpenSession();
             var query = session.CreateSQLQuery($"DELETE FROM Fhir_Diagnoses WHERE Id = '{diagnosis.Id}'");
             query.ExecuteUpdate();
 
-            query = session.CreateSQLQuery($"DELETE FROM Fhir_ConditionIcdTenCodes WHERE ConditionId = '{diagnosis.Condition.Id}'");
-            query.ExecuteUpdate();
+            if (hasIcdTen)
+            {
+                query = session.CreateSQLQuery($"DELETE FROM Fhir_ConditionIcdTenCodes WHERE ConditionId = '{diagnosis.Condition.Id}'");
+                query.ExecuteUpdate();
+            }
 
             query = session.CreateSQLQuery($"DELETE FROM Fhir_Conditions WHERE Id = '{diagnosis.Condition.Id}'");
             query.ExecuteUpdate();
